chatbot/requirements.txt
flask
flask-cors
pymupdf  # PDF text extraction
openai  # AI Q&A
python-dotenv